Submit Lab

+++ Lab1/src/Lab1/Lab1_2.java
@@ -4,8 +4,8 @@ public class Lab1_2 {
 
 	public static void main(String[] args) {
 		Scanner input = new Scanner(System.in);
-		double W = input.nextInt();
-		double H = input.nextInt();
+		double W = input.nextDouble();
+		double H = input.nextDouble();
 		double Area = W * H;
 		System.out.println(Area);
 		input.close();
